Check to see if server-side data is updated
Regardless of our TTL, we don't need to re-parse the data if the server-side data hasn't been updated. E.g., if someone only uses a 1 second TTL for the API, but Vatsim's server-side data only updates every 15 seconds. In that case, some of the calls would fetch the data but not parse and re-cache it